Connect
Connect button finds which boolean knobs (with filename) are selected, finds DeepRead with filename and either connects to DeepReformat under DeepRead or straight to DeepRead node depending on what is selected in the "ConnectTo:" pulldown choice. If "ConnectTo:" is set to "Reformat". Connect button checks first if DeepRead is connected to a Reformat node and if not, activates a warning message that points out which DeepRead isn't connected to a reformat. It finds if there are two or more DeepReads with same filename and only connects to one of them.